Revert "refactor(Select): move common styles to BaseSelect, reuse across select components"
This reverts commit d1957835b50f219ae0fc401075c6cff4913c3f17.
